requirements.txt
# NLP and LLM
langchain-ollama
python-dotenv

# Data Handling and Execution
pandas
sqlite3

# Frontend (for Vasavi)
streamlit

# Utilities
sqlparse